// src/app/client-view/models/house-hold-cash-flow.model.ts
export class HouseHoldCashFlow {
    id: string;
    name: string;
    lifeStyleExpensesFixed: number;
    lifeStyleExpensesDiscretionary: number;
    members: any;
    clientOnboardingStep: number;
    factFind: boolean;
    governmentBenefitIncome: number;
    primaryClientId: string;
    retirementYear: number;
    spouseClientId: string;
}